Renamed Home.Screen.tsx, Home.Style.tsx, and Home folder to MapHome.Screen.tsx, MapHome.Style.tsx, and MapHome to minimize confusion
