+++ Citizen-Science/screens/Mapp.Navigator.tsx
@@ -3,7 +3,7 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import {NavigationContainer} from '@react-navigation/native'
 import { FrontPageScreen } from '../screens/FrontPage/FrontPage.Screen';
 import {SafeAreaView} from 'react-native';
-import { HomeScreen } from '../screens/Home/Home.Screen';
+import { HomeScreen } from './MapHome/MapHome.Screen';
 import { MapScreen } from '../screens/Map/Map.Screen';
